# refactor(cli/restore): split backend into focused modules; keep `restore_cli.py` as CLI entry
## Summary
Modularize the restore CLI backend. `restore_cli.py` remains the top-level CLI with subcommands, while backend logic moves into dedicated files under `blase/cli/restore/`.

## Structure
- `blase/cli/restore/restore_cli.py`
  - Still the main entry point.
  - Keeps subcommands: `cmd_list`, `cmd_show`, `cmd_plan`, `cmd_run`.

- New modules
  - `anchors.py`: shared constants, type aliases, and symbolic anchors used across restore code.
  - `dispatcher.py`: orchestration entry points for CLI:
    - `run_step(...)`
    - `run_data(...)`
    - Primary targets invoked by `cmd_run`.
  - `pretty.py`: formatting utilities for plans, steps, and outputs.
  - `replay_cas.py`: CAS-related replay helpers and path/content checks.
  - `replay_exec.py`: step-level exec helpers called during plan replay.
    - Focused execution helpers used by `_exec_plan_for_step`.
  - `replay.py`:
    - `_build_stream_for_step(...)`
    - `_upstream_gen_for_sink(...)`
    - `_exec_plan_for_step(...)`

## Key changes
- Move execution and formatting logic out of `restore_cli.py` into small, testable modules.
- Centralize run dispatch in `dispatcher.py` so `cmd_run` has a single callsite.
- Keep CLI argument parsing and printing in `restore_cli.py` to preserve user-facing behavior.
- Isolate CAS ops and checks in `replay_cas.py`.
- Keep plan construction and upstream resolution in `replay.py`.
- Keep step-level execution helpers in `replay_exec.py` to limit cross-module coupling.
