+++ blase/cli/restore/restore_cli.py
@@ -1,15 +1,12 @@
-from __future__ import annotations
 import json
 import sqlite3
 from pathlib import Path
-from typing import Optional, Dict, Any, Tuple, Set, List
-import tempfile
+from typing import Optional, Dict, Any
 
-from blase.restore import step as restore_step
-from blase.restoring import materialize, store, cas, bindings, planner, code
-from blase.restoring.materialize import NeedReplay
-from blase.utils.config import RESTORE_CONFLICT as DEFAULT_CONFLICT, RESTORE_DEFAULT_DIR
-from blase.utils.hashing import Hash
+from blase.cli.restore.dispatcher import Policy, run_data, run_step
+from blase.cli.restore.pretty import _print_rows, _print_step
+from blase.restoring import cas
+from blase.utils.config import RESTORE_CONFLICT as DEFAULT_CONFLICT
 
 
 # --------- run discovery helpers ---------
@@ -80,199 +77,6 @@ def _open_db(run_path: Path) -> sqlite3.Connection:
     con = sqlite3.connect(db)
     con.row_factory = sqlite3.Row
     return con
-
-
-def _assert_path_matches_hash(path: Path, expected_hash: str, what: str) -> None:
-    got = Hash().hash_file(path)
-    if got != expected_hash:
-        raise SystemExit(
-            f"restore: {what} hash mismatch: got {got}, expected {expected_hash}"
-        )
-
-
-def _resolve_source_no_copy(
-    run_path: Path,
-    source_hash: str,
-    *,
-    seen_steps: Optional[Set[str]] = None,
-    created_paths: Optional[List[Path]] = None,
-) -> Tuple[Path, bool]:
-    """
-    Return a path to the recorded source without copying when possible.
-    If the original source_path exists and matches the recorded hash, use it.
-    Else replay/materialize (this will copy into /restored); we return that path and mark created=True.
-    """
-    # Try original source_of_truth path first (no copy)
-    src = store.load_data_source_path(run_path, source_hash)
-    if src:
-        p = Path(src)
-        if p.exists() and Hash().hash_file(p) == source_hash:
-            return p, False  # not created
-
-    # Fallback: ensure locally (may trigger replay of the producer; will copy/link into /restored)
-    p, created = _ensure_data_local_or_replay(
-        run_path,
-        source_hash,
-        store.get_data_kind(run_path, source_hash) or "csv",
-        seen_steps=seen_steps,
-        created_paths=created_paths,
-    )
-    # Track for cleanup if we had to copy
-    if created and created_paths is not None:
-        created_paths.append(p)
-    return p, created
-
-
-def _resolve_seed_no_copy_or_ephemeral(
-    run_path: Path,
-    seed_hash: str,
-    *,
-    seen_steps: Optional[Set[str]] = None,
-    created_paths: Optional[List[Path]] = None,
-) -> Tuple[Path, bool]:
-    """
-    Return a path to the recorded seed (run-1 output) without persisting it in /restored.
-    - If a valid existing path already exists (materialization or original), use it (no copy).
-    - Otherwise, replay the seed's producer into a *temporary* file (not /restored) and
-      do NOT record a materialization. Mark created=True and remember for cleanup.
-    """
-    # 1) Prefer an existing materialized path (no copy)
-    p = store.get_materialized_path(run_path, seed_hash)
-    if p and p.exists() and Hash().hash_file(p) == seed_hash:
-        return p, False
-
-    # 2) Try recorded source_path (the original file), if any
-    p2 = store.get_recorded_source_path(run_path, seed_hash)
-    if p2 and p2.exists() and Hash().hash_file(p2) == seed_hash:
-        return p2, False
-
-    # 3) No local copy → replay the seed into a temporary file (EPHEMERAL)
-    prod = store.producer_step_for_data(run_path, seed_hash)
-    if not prod:
-        raise SystemExit(
-            f"restore: seed {seed_hash} missing and no producer step recorded"
-        )
-
-    tmpdir = Path(tempfile.mkdtemp(prefix="blase-seed-"))
-    tmpfile = tmpdir / f"{seed_hash}.csv"
-
-    # Execute only the seed’s plan, writing to tmpfile; prevent materialization records.
-    _exec_plan_for_step(
-        run_path,
-        prod,
-        to_path=str(tmpfile),
-        backend_override=None,
-        seen_steps=seen_steps,
-        created_paths=None,
-        ephemeral_only=True,
-    )
-
-    if not tmpfile.exists() or Hash().hash_file(tmpfile) != seed_hash:
-        raise SystemExit("restore: ephemeral seed replay produced wrong bytes")
-
-    if created_paths is not None:
-        created_paths.append(tmpfile)
-
-    return tmpfile, True
-
-
-def _upstream_gen_for_sink(run_path: Path, target_step_hash: str):
-    """
-    Build an upstream (batch,last) generator for a Load.save_to_csv step by scanning
-    its plan backward to the nearest Transform.apply_function (preferred) or Extract.read_csv.
-    """
-    # figure out which sink we’re wiring
-    tip = store.load_step(run_path, target_step_hash)
-    tip_fqn = tip["function_fqn"]
-
-    # what counts as a streaming producer
-    def _is_extract_images(fqn: str) -> bool:
-        return fqn.endswith("Extract.read_images")
-
-    def _is_extract_csv(fqn: str) -> bool:
-        return fqn.endswith("Extract.read_csv")
-
-    def _is_extract_parquet(fqn: str) -> bool:
-        return fqn.endswith("Extract.read_parquet")
-
-    def _is_transform(fqn: str) -> bool:
-        return fqn.endswith("Transform.apply_function")
-
-    # ask the planner for a minimal chain and normalize it
-    raw_plan = planner.plan_for_step(run_path, target_step_hash) or []
-    nodes = _normalize_plan_nodes(run_path, raw_plan)
-
-    try:
-        idx = next(i for i, n in enumerate(nodes) if n["step_hash"] == target_step_hash)
-    except StopIteration:
-        raise SystemExit("restore: target sink step not found in plan")
-
-    # scan backwards to find a producer
-    upstream_node = None
-
-    # 1) nearest Transform.apply_function
-    for j in range(idx - 1, -1, -1):
-        if _is_transform(nodes[j]["function_fqn"]):
-            upstream_node = nodes[j]
-            break
-
-    # 2) if none, pick the right Extract fallback by sink type
-    if upstream_node is None:
-        if tip_fqn.endswith("Load.save_images_to_parquet"):
-            for j in range(idx - 1, -1, -1):
-                if _is_extract_images(nodes[j]["function_fqn"]):
-                    upstream_node = nodes[j]
-                    break
-                elif _is_extract_parquet(nodes[j]["function_fqn"]):
-                    upstream_node = nodes[j]
-                    break
-        else:
-            # default/legacy CSV path
-            for j in range(idx - 1, -1, -1):
-                if _is_extract_csv(nodes[j]["function_fqn"]):
-                    upstream_node = nodes[j]
-                    break
-
-    if upstream_node is None:
-        # helpful debug: show what we actually saw
-        dbg = " → ".join(f"{n['function_fqn']}[{n['step_hash'][:8]}]" for n in nodes)
-        raise SystemExit(
-            f"No upstream compute step found for sink replay.\n[debug] plan: {dbg}"
-        )
-
-    # delegate to the generic streaming builder for whatever node we found
-    return _build_stream_for_step(run_path, upstream_node["step_hash"])
-
-
-# --------- pretty helpers ---------
-
-
-def _print_rows(rows):
-    for r in rows:
-        print(" | ".join(str(v) for v in r))
-
-
-def _print_step(con, step_hash: str):
-    s = con.execute("SELECT * FROM steps WHERE step_hash=?", (step_hash,)).fetchone()
-    if not s:
-        print(f"[restore] step not found: {step_hash}")
-        return
-    print(f"step: {s['step_hash']}  fqn={s['function_fqn']}  status={s['status']}")
-    print(f"  run_id={s['run_id']}  ts_start={s['ts_start']}  ts_end={s['ts_end']}")
-    print(f"  params={s['params_json']}")
-    ins = con.execute(
-        "SELECT data_hash, role, arg_name FROM step_inputs WHERE step_hash=?",
-        (step_hash,),
-    ).fetchall()
-    print("  inputs:")
-    for r in ins:
-        print(f"    - {r['role']:8s} {r['data_hash']} arg={r['arg_name']}")
-    outs = con.execute(
-        "SELECT data_hash, name FROM step_outputs WHERE step_hash=?", (step_hash,)
-    ).fetchall()
-    print("  outputs:")
-    for r in outs:
-        print(f"    - {r['name']:8s} {r['data_hash']}")
 
 
 # --------- list/show/plan ---------
@@ -443,884 +247,6 @@ def cmd_plan(args):
 
 
 # --------- run (verify/materialize/replay) ---------
-
-
-def _restore_gen_for_manifest(run_path, manifest_hash, step_hash, store, bindings):
-    """
-    Build and run a restore generator for a recorded manifest.
-
-    The function inspects the upstream step to collect batch-identifiers, then
-    prefers a Parquet restore path if available, falling back to the images
-    restore path. It returns the generator produced by the selected bindings
-    call.
-
-    Parameters
-    ----------
-    run_path : pathlib.Path or str
-        Path to the run directory containing `nodes/nodes.db` and `cas/`.
-    manifest_hash : str
-        Data hash of the recorded table/images manifest to restore.
-    step_hash : str
-        Upstream step hash whose inputs encode batch descriptors.
-    store : object
-        Store API with `load_step_inputs(...)` and optionally
-        `read_parquet_params_for_manifest(...)` / `read_images_params_for_manifest(...)`.
-    bindings : object
-        Bindings API exposing `run_read_parquet_restore(...)` and
-        `run_read_images_restore(...)`.
-
-    Returns
-    -------
-    Iterator
-        Generator yielding 3-tuples ``(data, meta, is_last)`` where:
-        * `data` : backend-specific payload (e.g., tables or image batches)
-        * `meta` : dict with provenance (e.g., ordinal, batch descriptors)
-        * `is_last` : bool flagging the final batch
-
-    Raises
-    ------
-    SystemExit
-        If neither Parquet nor images restore parameters can be resolved
-        for the given `manifest_hash`.
-
-    Notes
-    -----
-    Parquet restore is attempted first. If the Parquet bindings raise an
-    exception, the function silently falls back to the images path.
-    """
-    ts_before = planner._step_row(planner._db(run_path), step_hash)["ts_start"]
-    ins = store.load_step_inputs(run_path, step_hash)
-
-    def _collect_batch_ids(ins):
-        roles = {"batch_desc", "table.batch.meta", "batch", "table.batch", "batchmeta"}
-        return [i["data_hash"] for i in ins if i.get("role") in roles]
-
-    batch_ids = _collect_batch_ids(ins)
-
-    # parquet path
-    rp_fn = getattr(store, "read_parquet_params_for_manifest", None)
-    if callable(rp_fn):
-        rp = rp_fn(run_path, manifest_hash, ts_before=ts_before)
-        if rp:
-            realized = {
-                "sources": rp.get("sources"),
-                "source": rp.get("source") or rp.get("directory"),
-                "manifest": manifest_hash,
-                "batch_descs": batch_ids,
-            }
-            try:
-                return bindings.run_read_parquet_restore(
-                    run_path=run_path, params=rp, realized=realized, transform_fn=None
-                )
-            except Exception:
-                pass  # fall back to images
-
-    # images path
-    ri_fn = getattr(store, "read_images_params_for_manifest", None)
-    if callable(ri_fn):
-        ri = ri_fn(run_path, manifest_hash, ts_before=ts_before)
-        if ri:
-            realized = {
-                "source": ri.get("directory"),
-                "manifest": manifest_hash,
-                "batch_descs": batch_ids,
-            }
-            return bindings.run_read_images_restore(
-                run_path=run_path, params=ri, realized=realized, transform_fn=None
-            )
-
-    raise SystemExit(
-        "replay: cannot find read_parquet or read_images params for manifest"
-    )
-
-
-def _exec_plan_for_step(
-    run_path: Path,
-    tip_step_hash: str,
-    to_path: Optional[str],
-    backend_override: Optional[str],
-    *,
-    seen_steps: Optional[Set[str]] = None,
-    created_paths: Optional[List[Path]] = None,
-    ephemeral_only: bool = False,
-) -> Dict[str, Path]:
-    """
-    Replay recorded steps up to ``tip_step_hash`` and materialize sinks.
-
-    Walks the stored plan in dependency order, reconstructs upstream generators
-    for extract steps, re-applies recorded transforms, and replays load steps to
-    produce the same on-disk artifacts (CSV files or Parquet shards). Supports
-    both images and table/parquet flows. For ``Transform.apply_function``,
-    prefers a Parquet-backed upstream when a manifest anchor is present, and
-    falls back to images if needed. When an upstream generator already exists,
-    it is piped directly into the next transform without re-anchoring.
-
-    Parameters
-    ----------
-    run_path : Path
-        Run root containing ``nodes/`` and ``cas/``.
-    tip_step_hash : str
-        Step hash to reach; all required upstream steps are replayed.
-    to_path : str or None
-        Optional final target override for the sink. When set, the last sink
-        writes here regardless of the recorded location.
-    backend_override : str or None
-        Optional sink backend override passed to replay handlers.
-    seen_steps : set of str, optional
-        Steps already executed in this process; they are skipped.
-    created_paths : list of Path, optional
-        Collector for all paths created during this replay.
-    ephemeral_only : bool, default False
-        If True, do not record materializations back to CAS.
-
-    Returns
-    -------
-    dict
-        Mapping ``{produced_data_hash: Path(output_file)}`` for each sink.
-
-    Raises
-    ------
-    SystemExit
-        If required anchors (e.g., manifest/source) are missing or if neither
-        Parquet nor images params can be resolved for a manifest-anchored
-        transform.
-
-    Notes
-    -----
-    Handles:
-      * ``Extract.read_csv`` → CSV stream restore.
-      * ``Extract.read_images`` → image batches via recorded manifest + descs.
-      * ``Extract.read_parquet`` → table/image-compatible batches with recorded
-        ``manifest`` and ``batch_desc_*``.
-      * ``Transform.apply_function`` → uses an existing upstream generator when
-        available; otherwise anchors by ``source`` or ``manifest``. Tries
-        ``read_parquet`` params first, then images.
-      * ``Load.save_to_csv`` and ``Load.save_images_to_parquet`` → write outputs,
-        validate expected hashes, and optionally record materializations.
-    """
-    plan = planner.plan_for_step(run_path, tip_step_hash) or []
-    upstream = None
-    produced: Dict[str, Path] = {}
-    produced_hashes: Set[str] = set()
-    seen_steps = seen_steps or set()
-    created_paths = created_paths if created_paths is not None else []
-
-    for sh in plan:
-        if sh in seen_steps:
-            continue
-        seen_steps.add(sh)
-
-        st = store.load_step(run_path, sh)
-        fqn = st["function_fqn"]
-
-        if fqn == "blase.Extract.read_csv":
-            ins = store.load_step_inputs(run_path, sh)
-            src_hash = next(
-                (i["data_hash"] for i in ins if i["role"] == "source"), None
-            )
-            realized = {}
-            if src_hash:
-                realized["__expected_source_hash__"] = src_hash
-            upstream = bindings.run_read_csv_restore(
-                run_path=run_path,
-                params=st["params"],
-                realized=realized,
-                transform_fn=None,
-            )
-            continue
-
-        if fqn == "blase.Extract.read_images":
-            ins = store.load_step_inputs(run_path, sh)  # code/env only
-            outs = store.load_step_outputs(run_path, sh)  # manifest + batch_desc_*
-
-            # recorded artifacts come from outputs
-            manifest_hash = next(
-                (o["data_hash"] for o in outs if o["name"] == "manifest"), None
-            )
-
-            def _ord(o):
-                n = o["name"]
-                try:
-                    return int(n.rsplit("_", 1)[-1])
-                except Exception:
-                    return 0
-
-            batch_descs = [
-                o["data_hash"]
-                for o in sorted(outs, key=_ord)
-                if o["name"].startswith("batch_desc_")
-            ]
-
-            realized = {}
-            if "directory" in st["params"]:
-                realized["source"] = st["params"]["directory"]
-            if manifest_hash:
-                realized["manifest"] = manifest_hash
-            if batch_descs:
-                realized["batch_descs"] = batch_descs
-
-            upstream = bindings.run_read_images_restore(
-                run_path=run_path,
-                params=st["params"],
-                realized=realized,
-                transform_fn=None,
-            )
-            continue
-
-        if fqn == "blase.Extract.read_parquet":
-            ins = store.load_step_inputs(run_path, sh)  # code/env only (unused here)
-            outs = store.load_step_outputs(run_path, sh)  # manifest + batch_desc_*
-
-            manifest_hash = next(
-                (o["data_hash"] for o in outs if o["name"] == "manifest"), None
-            )
-
-            def _ord(o):
-                n = o.get("name", "")
-                try:
-                    return int(n.rsplit("_", 1)[-1])
-                except Exception:
-                    return 0
-
-            batch_descs = [
-                o["data_hash"]
-                for o in sorted(outs, key=_ord)
-                if o.get("name", "").startswith("batch_desc_")
-            ]
-
-            realized = {}
-            # Prefer recorded sources list; fall back to single source if present.
-            if "sources" in st["params"]:
-                realized["sources"] = st["params"]["sources"]
-            elif "source" in st["params"]:
-                realized["source"] = st["params"]["source"]
-
-            if manifest_hash:
-                realized["manifest"] = manifest_hash
-            if batch_descs:
-                realized["batch_descs"] = batch_descs
-
-            upstream = bindings.run_read_parquet_restore(
-                run_path=run_path,
-                params=st["params"],
-                realized=realized,
-                transform_fn=None,
-            )
-            continue
-
-        if fqn == "blase.Transform.apply_function":
-            ins = store.load_step_inputs(run_path, sh)
-            fn = code.load_callable_from_blob(
-                cas.path_for(run_path, "code", store.pick_code_hash(ins))
-            )
-
-            # If an upstream generator already exists, consume it.
-            if upstream is not None:
-                upstream = bindings.run_apply_function_restore(
-                    run_path=run_path,
-                    params=st["params"],
-                    realized={"source_gen": upstream},
-                    transform_fn=fn,
-                )
-                continue
-
-            # Otherwise, anchor by manifest/source and build the upstream now.
-            def _pick_anchor(roles=("source", "manifest", "dataset", "manifest_root")):
-                for r in roles:
-                    h = next((i["data_hash"] for i in ins if i["role"] == r), None)
-                    if h:
-                        return r, h
-                return None, None
-
-            role, anchor_hash = _pick_anchor()
-            if not anchor_hash:
-                raise SystemExit(
-                    "replay: Transform.apply_function missing anchor (source|manifest|dataset)"
-                )
-
-            if role == "source":
-                src_path, _ = _resolve_source_no_copy(
-                    run_path,
-                    anchor_hash,
-                    seen_steps=seen_steps,
-                    created_paths=created_paths,
-                )
-                upstream = bindings.run_apply_function_restore(
-                    run_path=run_path,
-                    params=st["params"],
-                    realized={"source": src_path},
-                    transform_fn=fn,
-                )
-                continue
-
-            # treat anchor as a manifest; prefer parquet if present, else images
-            man_hash = anchor_hash
-            ts_before = planner._step_row(planner._db(run_path), sh)["ts_start"]
-
-            # Try parquet first
-            rp_params = getattr(
-                store, "read_parquet_params_for_manifest", lambda *a, **k: None
-            )(run_path, man_hash, ts_before=ts_before)
-            if rp_params:
-                ins_tr = store.load_step_inputs(run_path, sh)
-                batch_descs = [
-                    i["data_hash"]
-                    for i in ins_tr
-                    if i["role"]
-                    in (
-                        "batch_desc",
-                        "table.batch.meta",
-                        "batchmeta",
-                        "batch",
-                        "table.batch",
-                    )
-                ]
-                ins2 = store.load_step_inputs(run_path, sh)
-                manifest_hash = next(
-                    (i["data_hash"] for i in ins2 if i.get("role") == "manifest"), None
-                )
-                if not manifest_hash:
-                    raise SystemExit(
-                        "restore: cannot resolve manifest for upstream stream"
-                    )
-                gen = _restore_gen_for_manifest(
-                    run_path, manifest_hash, sh, store, bindings
-                )
-                upstream = bindings.run_apply_function_restore(
-                    run_path=run_path,
-                    params=st["params"],
-                    realized={"source_gen": gen},
-                    transform_fn=fn,
-                )
-                continue
-
-            # Fallback to images
-            ri_params = store.read_images_params_for_manifest(
-                run_path, man_hash, ts_before=ts_before
-            )
-            if not ri_params:
-                raise SystemExit(
-                    "replay: cannot find read_parquet or read_images params for manifest"
-                )
-
-            ins_tr = store.load_step_inputs(run_path, sh)
-            batch_descs = [
-                i["data_hash"]
-                for i in ins_tr
-                if i["role"]
-                in ("batch_desc", "table.batch.meta", "batch", "table.batch")
-            ]
-            gen = bindings.run_read_images_restore(
-                run_path=run_path,
-                params=ri_params,
-                realized={
-                    "source": ri_params.get("directory"),
-                    "manifest": man_hash,
-                    "batch": batch_descs,
-                },
-            )
-            upstream = bindings.run_apply_function_restore(
-                run_path=run_path,
-                params=st["params"],
-                realized={"source_gen": gen},
-                transform_fn=fn,
-            )
-            continue
-
-        if fqn == "blase.Load.save_to_csv":
-            ins = store.load_step_inputs(run_path, sh)
-            outs = store.load_step_outputs(run_path, sh)
-            expected_out_hash = outs[0]["data_hash"] if outs else None
-
-            if expected_out_hash and expected_out_hash in produced_hashes:
-                continue
-
-            # Pre-seed (append semantics)
-            preseed_path = None
-            seed_hash = _pick_viable_seed(run_path, ins)
-            if seed_hash:
-                preseed_path, created_seed = _resolve_seed_no_copy_or_ephemeral(
-                    run_path,
-                    seed_hash,
-                    seen_steps=seen_steps,
-                    created_paths=created_paths,
-                )
-                if created_seed and created_paths is not None:
-                    created_paths.append(preseed_path)
-
-            # Deterministic final target
-            if to_path:
-                target_override_this_sink = to_path
-                conflict_policy = "overwrite"
-            elif expected_out_hash:
-                target_override_this_sink = str(
-                    (RESTORE_DEFAULT_DIR / f"{expected_out_hash}.csv").resolve()
-                )
-                conflict_policy = "overwrite"
-            else:
-                default_name = Path(st["params"]["target"]).name
-                target_override_this_sink = str(
-                    (RESTORE_DEFAULT_DIR / default_name).resolve()
-                )
-                conflict_policy = "overwrite"
-
-            upstream_gen = upstream
-            if upstream_gen is None:
-                upstream_gen = _upstream_gen_for_sink(run_path, sh)
-
-            out = bindings.run_save_to_csv_replay(
-                run_path=run_path,
-                params=st["params"],
-                realized={},
-                upstream_gen=upstream_gen,
-                target_override=target_override_this_sink,
-                backend_override=backend_override,
-                preseed_path=preseed_path,
-                expected_out_hash=expected_out_hash,
-                on_conflict=conflict_policy,
-                record_materialization=(not ephemeral_only),
-            )
-            print(f"Replayed: {out}")
-
-            if expected_out_hash:
-                outp = Path(out)
-                produced[expected_out_hash] = outp
-                created_paths.append(outp)
-                produced_hashes.add(expected_out_hash)
-
-            upstream = None
-            continue
-
-        if fqn == "blase.Load.save_images_to_parquet":
-            ins = store.load_step_inputs(run_path, sh)
-            outs = store.load_step_outputs(run_path, sh)
-            st = store.load_step(run_path, sh)
-
-            def _ord(o):
-                n = o["name"]
-                return int(n.split("_")[-1]) if n.startswith("parquet_shard_") else 0
-
-            outs.sort(key=_ord)
-
-            expected_out_hashes = [
-                o["data_hash"] for o in outs if o["name"].startswith("parquet_shard")
-            ]
-
-            if to_path:
-                target_dir_override = to_path
-                conflict_policy = "overwrite"
-            else:
-                # default: deterministic directory under RESTORE_DEFAULT_DIR
-                # using the first expected hash as a hint to make the path unique.
-                RESTORE_DEFAULT_DIR.mkdir(parents=True, exist_ok=True)
-                target_dir_override = str((RESTORE_DEFAULT_DIR).resolve())
-                conflict_policy = "overwrite"
-
-            upstream_gen = upstream or _upstream_gen_for_sink(run_path, sh)
-            handler = bindings.RESTORE_HANDLERS[fqn]
-            out_paths = handler(
-                run_path=run_path,
-                params=st["params"],
-                realized={},
-                upstream_gen=upstream_gen,
-                target_override=target_dir_override,
-                on_conflict=conflict_policy,
-                expected_out_hashes=expected_out_hashes or None,
-                record_materialization=(not ephemeral_only),
-            )
-
-            print("Replayed shards:")
-            for p in out_paths:
-                print(f"  {p}")
-
-            # Track produced hashes → paths
-            # If outputs were recorded, map them 1:1; else compute here.
-            if outs:
-                for i, o in enumerate(outs):
-                    dh = o["data_hash"]
-                    if i < len(out_paths):
-                        produced[dh] = Path(out_paths[i])
-                        created_paths.append(Path(out_paths[i]))
-                        produced_hashes.add(dh)
-            else:
-                # No recorded outputs? compute and index
-                for p in out_paths:
-                    dh = Hash().hash_file(p)
-                    produced[dh] = Path(p)
-                    created_paths.append(Path(p))
-                    produced_hashes.add(dh)
-
-            upstream = None
-            continue
-
-        # Fallback for non-stream steps, if any
-        from blase import restore as _restore
-
-        _restore.step(run_path, sh)
-
-    return produced
-
-
-def _ensure_data_local_or_replay(
-    run_path: Path,
-    data_hash: str,
-    kind: str,
-    *,
-    seen_steps: Optional[Set[str]] = None,
-    created_paths: Optional[List[Path]] = None,
-) -> Tuple[Path, bool]:
-    """
-    Ensure a data hash is locally materialized. If not, replay its producer step
-    (using shared seen_steps/created_paths), then materialize it.
-    Returns (path, created_now).
-    """
-    if not kind:
-        kind = store.kind_for_hash(run_path, data_hash)
-    try:
-        ext = {
-            "parquet": ".parquet",
-            "parquet.shard": ".parquet",
-            "image.manifest": ".json",
-            "image.batch.meta": ".json",
-            "code": ".py",
-            "env": ".json",
-        }.get(kind, ".bin")
-        target_name = f"{data_hash}{ext}"
-        p = materialize.ensure_local(
-            run_path,
-            data_hash,
-            kind=kind,
-            policy="reuse",
-            to_dir=None,
-            target_name=target_name,
-        )
-        _assert_path_matches_hash(p, data_hash, f"{kind} materialization")
-        return p, False
-    except NeedReplay:
-        prod = store.producer_step_for_data(run_path, data_hash)
-        if not prod:
-            raise SystemExit(
-                f"restore: missing local copy and no producer for {data_hash}"
-            )
-
-        produced = _exec_plan_for_step(
-            run_path,
-            prod,
-            to_path=None,
-            backend_override=None,
-            seen_steps=seen_steps,
-            created_paths=created_paths,
-        )
-
-        p = produced.get(data_hash) if produced else None
-        if p is None:
-            # fallback: now that replay happened, ensure_local should succeed
-            p = materialize.ensure_local(
-                run_path, data_hash, kind=kind, policy="reuse", to_dir=None
-            )
-
-        _assert_path_matches_hash(p, data_hash, f"{kind} materialization after replay")
-        return p, True
-
-
-def _pick_viable_seed(run_path: Path, ins: list[dict]) -> Optional[str]:
-    """
-    From a list of inputs, pick a seed data hash that is either replayable (has a producer)
-    or already materialized. Skip dead seeds.
-    """
-    seeds = [i["data_hash"] for i in ins if i.get("role") == "seed"]
-    if not seeds:
-        return None
-    for dh in seeds:
-        if store.producer_step_for_data(run_path, dh):
-            return dh
-    for dh in seeds:
-        try:
-            kind = store.get_data_kind(run_path, dh) or "csv"
-            materialize.ensure_local(
-                run_path, dh, kind=kind, policy="reuse", to_dir=None
-            )
-            return dh
-        except Exception:
-            pass
-    return None
-
-
-def _normalize_plan_nodes(run_path: Path, plan):
-    """
-    Accept nodes as strings (step_hash), tuples (step_hash, ...),
-    or dicts; return a list of dicts with at least:
-      {"step_hash": <hash>, "function_fqn": <fqn>}
-    """
-    out = []
-    for node in plan:
-        if isinstance(node, dict):
-            if "step_hash" in node and "function_fqn" in node:
-                out.append(
-                    {
-                        "step_hash": node["step_hash"],
-                        "function_fqn": node["function_fqn"],
-                    }
-                )
-            elif "hash" in node:
-                st = store.load_step(run_path, node["hash"])
-                out.append(
-                    {"step_hash": node["hash"], "function_fqn": st["function_fqn"]}
-                )
-            else:
-                # Last resort: try to find something that looks like a hash
-                h = (
-                    node.get("id")
-                    or node.get("step")
-                    or node.get("node")
-                    or node.get("sha")
-                )
-                if not h:
-                    continue
-                st = store.load_step(run_path, h)
-                out.append({"step_hash": h, "function_fqn": st["function_fqn"]})
-        elif isinstance(node, (list, tuple)) and node:
-            h = node[0]
-            st = store.load_step(run_path, h)
-            out.append({"step_hash": h, "function_fqn": st["function_fqn"]})
-        elif isinstance(node, str):
-            h = node
-            st = store.load_step(run_path, h)
-            out.append({"step_hash": h, "function_fqn": st["function_fqn"]})
-    return out
-
-
-STREAM_FQNS = (
-    "blase.Extract.read_csv",
-    "blase.Extract.read_images",
-    "blase.Extract.read_parquet",
-    "blase.Transform.apply_function",
-)
-
-
-def _is_stream_fqn(fqn: str) -> bool:
-    return any(fqn.endswith(s) for s in STREAM_FQNS)
-
-
-def _build_stream_for_step(run_path: Path, step_hash: str):
-    """
-    Construct a streaming generator for a recorded step.
-
-    Given a `step_hash`, this inspects the recorded function type and returns
-    an upstream generator compatible with replay bindings:
-    - `Extract.read_csv` → `bindings.run_read_csv_restore(...)`
-    - `Extract.read_images` → `bindings.run_read_images_restore(...)`
-    - `Transform.apply_function` → chains to the nearest upstream stream
-      (from the plan or via lineage), then returns
-      `bindings.run_apply_function_restore(...)` seeded by that stream.
-    Non-stream steps delegate to `blase.restore.step(..., kind="data")`.
-
-    Parameters
-    ----------
-    run_path : Path
-        Root directory of the recorded run.
-    step_hash : str
-        Hash of the step to materialize as a stream.
-
-    Returns
-    -------
-    iterator
-        A generator yielding `(batch, is_last)` tuples, or a binding-specific
-        generator that internally yields `(batch, meta, is_last)` for image
-        pipelines and is consumed by downstream replay code. For non-stream
-        steps, the return value of `blase.restore.step(..., kind="data")` is
-        forwarded.
-
-    Notes
-    -----
-    - For `Extract.read_csv`, the realized inputs include
-      `__expected_source_hash__` when a recorded source hash exists.
-    - For `Extract.read_images`, realized inputs may include `manifest`
-      and a list of `batch` hashes when present in recorded inputs.
-    - For `Transform.apply_function`, the function:
-        1) Attempts to find the nearest upstream streaming node from the
-           topological plan. If none is found, it falls back to lineage:
-           manifest → batch_desc → batch producer step.
-        2) Loads the exact recorded callable via `code.load_callable_from_blob`
-           using the recorded code hash.
-        3) Returns the binding `run_apply_function_restore` wired to the
-           upstream stream (or a materialized path for older CSV runs).
-    - If no suitable upstream can be located for a transform, a
-      `SystemExit` is raised by the caller block that invokes this helper.
-
-    See Also
-    --------
-    run_read_csv_restore
-    run_read_images_restore
-    run_apply_function_restore
-    blase.restore.step
-    """
-    st = store.load_step(run_path, step_hash)
-    fqn = st["function_fqn"]
-
-    # ---- Extract.read_csv ----
-    if fqn.endswith("Extract.read_csv"):
-        ins = store.load_step_inputs(run_path, step_hash)
-        realized = {}
-        src_hash = next((i["data_hash"] for i in ins if i["role"] == "source"), None)
-        if src_hash:
-            realized["__expected_source_hash__"] = src_hash
-        return bindings.run_read_csv_restore(
-            run_path=run_path, params=st["params"], realized=realized, transform_fn=None
-        )
-
-    # ---- Extract.read_images ----
-    if fqn.endswith("Extract.read_images"):
-        ins = store.load_step_inputs(run_path, step_hash)
-        manifest_hash = next(
-            (i["data_hash"] for i in ins if i["role"] == "manifest"), None
-        )
-        batch_hashes = [i["data_hash"] for i in ins if i["role"] == "batch"]
-        realized = {"source": st["params"].get("directory")}
-        if manifest_hash:
-            realized["manifest"] = manifest_hash
-        if batch_hashes:
-            realized["batch"] = batch_hashes
-        return bindings.run_read_images_restore(
-            run_path=run_path, params=st["params"], realized=realized, transform_fn=None
-        )
-
-    # ---- Extract.read_parquet ----
-    if fqn.endswith("Extract.read_parquet"):
-        ins = store.load_step_inputs(run_path, step_hash)
-        manifest_hash = next(
-            (i["data_hash"] for i in ins if i["role"] == "manifest"), None
-        )
-        batch_hashes = [i["data_hash"] for i in ins if i["role"] == "batch"]
-
-        realized = {}
-        if "sources" in st["params"]:
-            realized["sources"] = st["params"]["sources"]
-        elif "source" in st["params"]:
-            realized["source"] = st["params"]["source"]
-
-        if manifest_hash:
-            realized["manifest"] = manifest_hash
-        if batch_hashes:
-            realized["batch"] = batch_hashes
-
-        return bindings.run_read_parquet_restore(
-            run_path=run_path,
-            params=st["params"],
-            realized=realized,
-            transform_fn=None,
-        )
-
-    # ---- Transform.apply_function (chain to its immediate upstream) ----
-    if fqn.endswith("Transform.apply_function"):
-        raw_plan = planner.plan_for_step(run_path, step_hash) or []
-        nodes = _normalize_plan_nodes(run_path, raw_plan)
-        try:
-            idx = next(i for i, n in enumerate(nodes) if n["step_hash"] == step_hash)
-        except StopIteration:
-            raise SystemExit("restore: transform step not found in plan")
-        upstream_node = None
-        for j in range(idx - 1, -1, -1):
-            if _is_stream_fqn(nodes[j]["function_fqn"]):
-                upstream_node = nodes[j]
-                break
-        if upstream_node is None:
-            # ---- Resolve by lineage hashes (manifest / batch_desc / batch) ----
-            ins = store.load_step_inputs(run_path, step_hash)
-
-            # prefer a manifest hash if present
-            man = next((i["data_hash"] for i in ins if i["role"] == "manifest"), None)
-            if man:
-                ex = store.producer_step_for_data(run_path, man)
-                if ex:
-                    upstream_node = {
-                        "step_hash": ex,
-                        "function_fqn": store.load_step(run_path, ex)["function_fqn"],
-                    }
-
-            # else try batch_desc, then batch
-            if upstream_node is None:
-                bdesc = next(
-                    (
-                        i["data_hash"]
-                        for i in ins
-                        if i["role"]
-                        in (
-                            "batch_desc",
-                            "batch_meta",
-                            "batchmeta",
-                            "batch_desc_1",
-                            "batch_desc_2",
-                            "table.batch.meta",
-                        )
-                    ),
-                    None,
-                )
-                if bdesc:
-                    ex = store.producer_step_for_data(run_path, bdesc)
-                    if ex:
-                        upstream_node = {
-                            "step_hash": ex,
-                            "function_fqn": store.load_step(run_path, ex)[
-                                "function_fqn"
-                            ],
-                        }
-
-            if upstream_node is None:
-                b = next(
-                    (
-                        i["data_hash"]
-                        for i in ins
-                        if i["role"] in ("batch", "image.batch", "table.batch")
-                    ),
-                    None,
-                )
-                if b:
-                    ex = store.producer_step_for_data(run_path, b)
-                    if ex:
-                        upstream_node = {
-                            "step_hash": ex,
-                            "function_fqn": store.load_step(run_path, ex)[
-                                "function_fqn"
-                            ],
-                        }
-        if upstream_node is None:
-            # Fallback: use recorded CSV source path if present (older runs)
-            ins = store.load_step_inputs(run_path, step_hash)
-            src = next((i["data_hash"] for i in ins if i["role"] == "source"), None)
-            if not src:  # nothing to chain
-                raise SystemExit(
-                    "restore: cannot locate upstream producer for Transform.apply_function"
-                )
-            p = store.get_materialized_path(
-                run_path, src
-            ) or store.get_recorded_source_path(run_path, src)
-            if not p or not p.exists():
-                p = materialize.ensure_local(
-                    run_path, src, kind=store.get_data_kind(run_path, src)
-                )
-            upstream_gen = p.as_posix()
-        else:
-            upstream_gen = _build_stream_for_step(run_path, upstream_node["step_hash"])
-
-        # load the exact user callable
-        ins = store.load_step_inputs(run_path, step_hash)
-        code_hash = store.pick_code_hash(ins)
-        fn = code.load_callable_from_blob(cas.path_for(run_path, "code", code_hash))
-
-        return bindings.run_apply_function_restore(
-            run_path=run_path,
-            params=st["params"],
-            realized={"source": upstream_gen},
-            transform_fn=fn,
-        )
-
-    # Non-stream fallback
-    from blase import restore as restore_mod
-
-    return restore_mod.step(run_path, step_hash, kind="data")
-
-
 def cmd_run(args):
     """
     Run the `blase restore run` CLI for a data hash or a step hash.
@@ -1405,252 +331,28 @@ def cmd_run(args):
     if not getattr(args, "on_conflict", None):
         args.on_conflict = DEFAULT_CONFLICT
 
-    # ---------------------------
-    # DATA-CENTRIC RESTORE PATH
-    # ---------------------------
-    if getattr(args, "data", None):
-        data_hash = args.data
-        kind = store.get_data_kind(run_path, data_hash) or "data"
-
-        # try fast materialize first
-        try:
-            if args.to:
-                to_path = Path(args.to).resolve()
-                out = materialize.ensure_local(
-                    run_path,
-                    data_hash,
-                    kind=kind,
-                    to_dir=to_path.parent,
-                    target_name=to_path.name,
-                    on_conflict=args.on_conflict,
-                )
-            else:
-                RESTORE_DEFAULT_DIR.mkdir(parents=True, exist_ok=True)
-                ext = "csv" if kind == "csv" else "bin"
-                out = materialize.ensure_local(
-                    run_path,
-                    data_hash,
-                    kind=kind,
-                    to_dir=RESTORE_DEFAULT_DIR,
-                    target_name=f"{data_hash}.{ext}",
-                    on_conflict=args.on_conflict,
-                )
-            _assert_path_matches_hash(Path(out), data_hash, "fast-path materialize")
-            print(f"Materialized: {out}")
-            return 0
-        except NeedReplay:
-            pass  # fall through to replay
-
-        # Replay path
-        seen_steps: Set[str] = set()
-        created_paths: List[Path] = []
-
-        # find the producer step and execute its forward plan
-        prod = store.producer_step_for_data(run_path, data_hash)
-        if not prod:
-            raise SystemExit(
-                f"[restore] no producer step recorded for data {data_hash}"
-            )
-
-        produced = _exec_plan_for_step(
-            run_path,
-            prod,
-            to_path=args.to,
-            backend_override=getattr(args, "backend", None),
-            seen_steps=seen_steps,
-            created_paths=created_paths,
-        )
-
-        # ---- VERIFY FINAL ARTIFACT MATCHES REQUESTED HASH ----
-        # 1) Prefer the actual path that produced THIS hash (robust against renames)
-        final_path = produced.get(data_hash)
-
-        # 2) Fallbacks if mapping isn’t available (shouldn’t happen, but be safe)
-        if final_path is None:
-            if args.to:
-                final_path = Path(args.to).resolve()
-            else:
-                # fallback to sink target basename under RESTORE_DEFAULT_DIR
-                sink_step = store.producer_step_for_data(run_path, data_hash)
-                if not sink_step:
-                    raise SystemExit(
-                        "restore: cannot locate producing sink for final verification"
-                    )
-                sink_meta = store.load_step(run_path, sink_step)
-                default_name = Path(
-                    sink_meta["params"].get("target", f"{data_hash}.csv")
-                ).name
-                final_path = (RESTORE_DEFAULT_DIR / default_name).resolve()
-
-        if not final_path.exists():
-            raise SystemExit(
-                f"restore: expected final output does not exist: {final_path}"
-            )
-
-        _assert_path_matches_hash(final_path, data_hash, "final output")
-        print(f"Verified: {final_path} == {data_hash[:16]}…")
-
-        # --- Cleanup intermediates unless user asked to keep them ---
-        if not getattr(args, "keep_intermediates", False):
-            final_real = final_path.resolve()
-
-            # De-dupe by resolved path and remove everything except the final
-            seen: set[Path] = set()
-            for p in created_paths:
-                try:
-                    r = p.resolve()
-                except Exception:
-                    continue
-                if r in seen:
-                    continue
-                seen.add(r)
-                if r == final_real:
-                    continue
-                try:
-                    Path(r).unlink(missing_ok=True)
-                except Exception:
-                    # best-effort: ignore failures (e.g., already deleted)
-                    pass
-
-            return 0
-
-        # Non-sink producer: stream verify only for now
-        if args.mode in ("verify",):
-            gen = (
-                _build_stream_for_step(run_path, prod)
-                if _is_stream_fqn(store.load_step(run_path, prod)["function_fqn"])
-                else restore_step(run_path, prod, kind="data")
-            )
-            total = 0
-            for i, (b, last) in enumerate(gen, 1):
-                n = len(b) if hasattr(b, "__len__") else "?"
-                print(f"[verify] batch {i}: {n} rows  last={last}")
-                if isinstance(n, int):
-                    total += n
-                if getattr(args, "limit_batches", None) and i >= args.limit_batches:
-                    break
-            print(f"[verify] total rows (best-effort): {total}")
-            return 0
-
-        raise SystemExit(
-            "restore: replay for non-sink data is not implemented yet in --data mode. "
-            "Use `--step <producer_step>` or `--mode verify`."
-        )
-
-    # ---------------------------
-    # STEP-CENTRIC RESTORE PATH
-    # ---------------------------
-    if not getattr(args, "step", None):
-        raise SystemExit(
-            "restore: either --step or --data is required (mutually exclusive)."
-        )
-
-    step_hash = args.step
-
-    # Load step metadata up front (decide sink vs non-sink)
-    st = store.load_step(run_path, step_hash)
-    fqn = st["function_fqn"]
-
-    # Wrapper for "verify" and "replay"
-    limit_batches = getattr(args, "limit_baches", None)
-
-    # ---- MATERIALIZE: just bring recorded outputs back (no replay) ----
-    if args.mode == "materialize":
-        outs = store.load_step_outputs(run_path, step_hash)
-        if not outs:
-            print("No outputs recorded for this step.")
-            return 2
-
-        # Take the first output
-        out_hash = outs[0]["data_hash"]
-        kind = store.get_data_kind(run_path, out_hash) or "data"
-
-        # Resolve target path pieces
-        to_path = Path(args.to).resolve()
-        to_dir = to_path.parent
-        target_name = to_path.name
-
-        try:
-            path = materialize.ensure_local(
-                run_path,
-                out_hash,
-                kind=kind,
-                to_dir=to_dir,
-                target_name=target_name,
-                on_conflict=args.on_conflict,
-            )
-            print(f"Materialized: {path}")
-            return 0
-        except materialize.NeedReplay:
-            print("Output not available locally. Use:")
-            print(f"  blase restore plan --step {step_hash}")
-            print(f"  blase restore run  --step {step_hash} --mode replay")
-            return 3
-
-    # ---- VERIFY: stream results without writing ----
-    if args.mode == "verify":
-        if fqn in ("blase.Load.save_to_csv", "blase.Load.save_images_to_parquet"):
-            upstream_gen = _upstream_gen_for_sink(run_path, step_hash)
-            gen = upstream_gen
-        elif _is_stream_fqn(fqn):
-            gen = _build_stream_for_step(run_path, step_hash)
+    has_step = bool(getattr(args, "step", None))
+    has_data = bool(getattr(args, "data", None))
+    mode = getattr(args, "mode", None)
+    if mode is None:
+        if has_step:
+            mode = "replay"
+        elif has_data:
+            mode = "materialize"
         else:
-            gen = restore_step(run_path, step_hash, kind="data")
-
-        def _coerce_stream_2(gen):
-            for item in gen:
-                if isinstance(item, tuple):
-                    if len(item) == 3:
-                        b, _meta, last = item
-                    elif len(item) == 2:
-                        b, last = item
-                    else:
-                        b, last = item, False
-                else:
-                    b, last = item, False
-                yield b, last
-
-        total = 0
-        for i, (b, last) in enumerate(_coerce_stream_2(gen), 1):
-            try:
-                n = len(b)
-            except Exception:
-                n = "?"
-            print(f"[verify] batch {i}: {n} rows  last={last}")
-            if isinstance(n, int):
-                total += n
-            if limit_batches and i >= limit_batches:
-                break
-        print(f"[verify] total rows (best-effort): {total}")
-        return 0
-
-    # ---- REPLAY: for sinks, wire upstream; for non-sinks, just restore ----
-    if args.mode == "replay":
-        if fqn in ("blase.Load.save_to_csv", "blase.Load.save_images_to_parquet"):
-            upstream_gen = _upstream_gen_for_sink(run_path, step_hash)
-            handler = bindings.RESTORE_HANDLERS[fqn]  # handler must exist for both FQNs
-            out_path = handler(
-                run_path=run_path,
-                params=st["params"],
-                realized={},  # sinks don’t need extra realized inputs
-                upstream_gen=upstream_gen,
-                target_override=getattr(args, "to", None),
-                backend_override=getattr(args, "backend", None),
+            raise SystemExit(
+                "restore: either --step or --data is required (mutually exclusive)."
             )
-            print(out_path)
-            return 0
 
-        # non-sinks: as before
-        gen = (
-            _build_stream_for_step(run_path, step_hash)
-            if _is_stream_fqn(fqn)
-            else restore_step(run_path, step_hash, kind="data")
-        )
-        for i, (b, last) in enumerate(gen, 1):
-            n = len(b) if hasattr(b, "__len__") else "?"
-            print(f"batch {i}: {n} items, last={last}")
-            if getattr(args, "limit_batches", None) and i >= args.limit_batches:
-                break
-        return 0
+    pol = Policy(
+        on_conflict=args.on_conflict,
+        keep_intermediates=getattr(args, "keep_intermediates", False),
+        limit_batches=getattr(args, "limit_batches", None),
+        backend=getattr(args, "backend", None),
+        mode=mode,
+        to=getattr(args, "to", None),
+    )
 
-    raise SystemExit(f"[restore] unknown mode: {args.mode}")
+    if has_step:
+        return run_step(run_path, args.step, mode, pol.to, pol)
+    return run_data(run_path, args.data, mode, pol.to, pol)
